refactor tests with an acceptance helper
