+++ .eslintrc.js
@@ -8,6 +8,5 @@ module.exports = {
   env: {
     browser: true
   },
-  rules: {
-  }
+  rules: {}
 };
